+++ src/ElasticLogger.Test/ElasticsearchLoggerFilterTests.cs
@@ -1,0 +1,302 @@
+﻿using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
+using AM.Extensions.Logging.ElasticSearch;
+using System;
+using Xunit;
+using Microsoft.Extensions.Configuration.Memory;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using Nest;
+using Xunit.Abstractions;
+
+namespace ElasticLogger.Test
+{
+    public class ElasticsearchLoggerFilterTests : IClassFixture<ESFixture>
+    {
+        private readonly ITestOutputHelper _output;
+        private readonly ESFixture _fixture;
+
+        public ElasticsearchLoggerFilterTests(ESFixture fixture, ITestOutputHelper output)
+        {
+            _fixture = fixture;
+            _output = output;
+        }
+
+        [Fact]
+        public async Task No_categories_for_es_should_use_default()
+        {
+            await _fixture.ReadyAsync();
+
+
+            var config = new ConfigurationBuilder()
+                .Add(new MemoryConfigurationSource
+                {
+                    InitialData = new Dictionary<string, string>
+                    {
+                            {"Logging:Elasticsearch:IncludeScopes", "true"},
+                            {"Logging:Elasticsearch:LogLevel:Default", "Error"},
+                            {"Logging:Elasticsearch:LogLevel:My.Bananas.Trace", "Trace"},
+                            {"Logging:LogLevel:Default", "Information"}
+                    }
+                })
+                .Build();
+
+            var serviceProvider = new ServiceCollection()
+                .AddLogging(x =>
+                {
+                    x.AddConfiguration(config.GetSection("Logging"));
+                    x.AddElasticSearch(options =>
+                    {
+                        options.ElasticsearchEndpoint = _fixture.Endpoint;
+                        options.IndexName = "trace";
+                    });
+                })
+                .BuildServiceProvider();
+
+            var loggerFactory = serviceProvider.GetService<ILoggerFactory>();
+            var logger = loggerFactory.CreateLogger("My.Bananas.Trace");
+
+            logger.LogTrace("bananas taste yucky");
+
+            var delayTask = Task.Delay(TimeSpan.FromSeconds(2));
+            var client = new ElasticClient(new ConnectionSettings(_fixture.Endpoint));
+            await client.PingAsync();
+            await delayTask;
+
+            var resp = await client.CatIndicesAsync();
+
+            Assert.Single(resp.Records);
+
+            var dyndocs = await client.SearchAsync<dynamic>(s => s
+                //.Index("trace-*")
+                .AllIndices()
+                .AllTypes());
+
+            Assert.Single(dyndocs.Documents);
+
+        }
+
+
+        [Fact]
+        public async Task configured_categories_dont_log_for_config()
+        {
+
+            var categoryName = "Billy.bob";
+
+
+            await _fixture.ReadyAsync();
+
+            var config = new ConfigurationBuilder()
+                .Add(new MemoryConfigurationSource
+                {
+                    InitialData = new Dictionary<string, string>
+                    {
+                        {"Logging:Elasticsearch:IncludeScopes", "true"},
+                        {"Logging:Elasticsearch:LogLevel:Default", "Error"},
+                        {"Logging:Elasticsearch:LogLevel:" + categoryName, "Error"},
+                        {"Logging:LogLevel:Default", "Error"}
+                    }
+                })
+                .Build();
+
+            var serviceProvider = new ServiceCollection()
+                .AddLogging(x =>
+                {
+                    x.AddConfiguration(config.GetSection("Logging"));
+                    x.AddElasticSearch(options =>
+                    {
+                        options.ElasticsearchEndpoint = _fixture.Endpoint;
+                        options.IndexName = "trace";
+                    });
+                })
+                .BuildServiceProvider();
+
+            var loggerFactory = serviceProvider.GetService<ILoggerFactory>();
+            var logger = loggerFactory.CreateLogger(categoryName);
+
+            logger.LogTrace("bananas taste yucky");
+
+            var delayTask = Task.Delay(TimeSpan.FromSeconds(5));
+            var client = new ElasticClient(new ConnectionSettings(_fixture.Endpoint));
+            await client.PingAsync();
+            await delayTask;
+
+            var resp = await client.CatIndicesAsync();
+
+            Assert.Empty(resp.Records);
+
+
+            var dyndocs = await client.SearchAsync<dynamic>(s => s
+                //.Index("trace-*")
+                .AllIndices()
+                .AllTypes());
+
+            Assert.Single(dyndocs.Documents);
+
+        }
+
+
+        [Fact]
+        public async Task No_elasticsearch_section_should_use_default_log_level()
+        {
+            await _fixture.ReadyAsync();
+
+            var config = new ConfigurationBuilder()
+                            .Add(new MemoryConfigurationSource
+                            {
+                                InitialData = new Dictionary<string, string>
+                                {
+                                    {"Logging:LogLevel:Default", "Trace" }
+                                }
+                            })
+                            .Build();
+
+
+            var serviceProvider = new ServiceCollection()
+                .AddLogging(x =>
+                {
+                    x.AddConfiguration(config.GetSection("Logging"));
+                    x.AddElasticSearch(options =>
+                    {
+                        options.ElasticsearchEndpoint = _fixture.Endpoint;
+                        options.IndexName = "trace";
+                    });
+                })
+                .BuildServiceProvider();
+
+            var loggerFactory = serviceProvider.GetService<ILoggerFactory>();
+            var logger = loggerFactory.CreateLogger("My.Bananas.Trace");
+
+            logger.LogTrace("bananas taste yucky");
+
+            var delayTask = Task.Delay(TimeSpan.FromSeconds(5));
+            var client = new ElasticClient(new ConnectionSettings(_fixture.Endpoint));
+            await client.PingAsync();
+            await delayTask;
+
+            var resp = await client.CatIndicesAsync();
+
+            Assert.Single(resp.Records);
+
+            var inm = new IndexName();
+            inm.Name = "trace-*";
+            var i = Indices.Index(inm);
+
+            var indxResponse = await client.GetIndexAsync(Indices.Index("trace-*"));
+
+            Assert.Single(indxResponse.Indices);
+
+            //s=>s.Index(Indices.AllIndices).
+
+            //var tester = client.SearchAsync<dynamic>(s => s.From(0).Size(1)
+            //    .Query(q => q
+            //        .Terms(t => t
+            //            .Name("named_query")
+            //            .Boost(1.1f)
+            //            .Field("cvrNummer")
+            //            .Terms("36406208"))));
+
+
+            var docs = await client.SearchAsync<dynamic>(s => s
+                .AllIndices()
+                .From(0)
+                .Size(100));
+            //.Query(q=>q.Match(m=>m.Field(f=>f)))
+
+            Assert.Single(docs.Documents);
+
+
+        }
+
+
+        [Fact]
+        public async Task FullESWriteTest()
+        {
+            await _fixture.ReadyAsync();
+
+            ILoggerFactory loggerFactory = new LoggerFactory()
+                .AddElasticSearch(_fixture.Endpoint);
+
+            var logger = loggerFactory.CreateLogger("xxxxxxx");
+
+            var circularRefObj = new Circle();
+            circularRefObj.me = circularRefObj;
+
+            logger.Log(Microsoft.Extensions.Logging.LogLevel.Critical, new EventId(), circularRefObj, null, (circle, exception) => "");
+
+            var delayTask = Task.Delay(TimeSpan.FromSeconds(5));
+            var client = new ElasticClient(new ConnectionSettings(_fixture.Endpoint));
+            await client.PingAsync();
+            await delayTask;
+
+            var resp = await client.CatIndicesAsync();
+
+            Assert.Single(resp.Records);
+
+
+            var dyndocs = await client.SearchAsync<dynamic>(s => s
+                //.Index("trace-*")
+                .AllIndices()
+                .AllTypes());
+
+            Assert.Single(dyndocs.Documents);
+        }
+
+
+
+
+        [Fact]
+        public async Task FullSimpleESTest()
+        {
+                await _fixture.ReadyAsync();
+                var client = new ElasticClient(new ConnectionSettings(_fixture.Endpoint));
+                await client.PingAsync();
+
+                var tweet = new Tweet
+                {
+                    Id = 1,
+                    User = "kimchy",
+                    PostDate = new DateTime(2009, 11, 15),
+                    Message = "Trying out NEST, so far so good?"
+                };
+                var response = await client.IndexAsync(tweet, idx => idx.Index("mytweetindex"));
+                var response2 = client.Get<Tweet>(1, idx => idx.Index("mytweetindex"));
+                var tweetResp = response;
+                var tweetResp2 = response2.Source;
+                Assert.Equal(tweet, tweetResp2);
+
+                await Task.Delay(TimeSpan.FromSeconds(5));
+
+                var dyndocs = await client.SearchAsync<dynamic>(s => s
+                    .Index("mytweetindex")
+                    .AllTypes());
+                //.AllIndices());
+                //.From(0)
+                //.Size(100)
+                //.Query(q => q.
+                //    Match(m => m.
+                //        Field("User")
+                //        .Query("kimchy"))));
+
+                Assert.Single(dyndocs.Documents);
+
+
+                var docs = await client.SearchAsync<Tweet>(s => s
+                    .Index("mytweetindex")
+                    .AllTypes());
+                //.AllIndices()
+                //.From(0)
+                //.Size(100)
+                //.Query(q => q.
+                //    Match(m => m.
+                //        Field("User")
+                //        .Query("kimchy"))));
+
+                Assert.Single(docs.Documents);
+        }
+
+
+
+    }
+}
